browser/components/backup/tests/chrome/test_disable_backup_encryption.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Tests for the disable-backup-encryption component</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script
  src="chrome://browser/content/backup/disable-backup-encryption.mjs"
  type="module"
></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script>

    const { BrowserTestUtils } = ChromeUtils.importESModule(
      "resource://testing-common/BrowserTestUtils.sys.mjs"
    );

    /**
     * Tests that adding a disable-backup-encryption element to the DOM causes it to
     * fire a BackupUI:InitWidget event.
     */
    add_task(async function test_initWidget() {
      let disableBackupEncryption = document.createElement("disable-backup-encryption");
      let content = document.getElementById("content");

      let sawInitWidget = BrowserTestUtils.waitForEvent(content, "BackupUI:InitWidget");
      content.appendChild(disableBackupEncryption);
      await sawInitWidget;
      ok(true, "Saw BackupUI:InitWidget");

      disableBackupEncryption.remove();
    });

    /**
     * Tests that pressing the confirm button will dispatch the expected events.
     */
    add_task(async function test_confirm() {
      let disableBackupEncryption = document.getElementById("test-disable-backup-encryption");
      let confirmButton = disableBackupEncryption.confirmButtonEl;

      ok(confirmButton, "Confirm button should be found");

      let content = document.getElementById("content");
      let promise = BrowserTestUtils.waitForEvent(content, "disableEncryption");

      confirmButton.click()

      await promise;
      ok(true, "Detected event after selecting the confirm button");
    })

    /**
     * Tests that pressing the cancel button will dispatch the expected events.
     */
    add_task(async function test_cancel() {
      let disableBackupEncryption = document.getElementById("test-disable-backup-encryption");
      let cancelButton = disableBackupEncryption.cancelButtonEl;

      ok(cancelButton, "Cancel button should be found");

      let content = document.getElementById("content");
      let promise = BrowserTestUtils.waitForEvent(content, "dialogCancel");

      cancelButton.click()

      await promise;
      ok(true, "Detected event after selecting the cancel button");
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  <disable-backup-encryption id="test-disable-backup-encryption"></disable-backup-encryption>
</div>
<pre id="test"></pre>
</body>
</html>
